Refactor CLI container commands
